pyproject.toml
[project]
name = "x-dm-moderator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "==3.12.12"
dependencies = [
    "aiosqlite>=0.22.1",
    "fastapi>=0.135.1",
    "httpx>=0.28.1",
    "pydantic>=2.12.0",
    "pyyaml>=6.0.3",
    "torch>=2.10.0",
    "transformers>=5.3.0",
    "uvicorn>=0.41.0",
]

[dependency-groups]
dev = [
    "prek>=0.3.5",
    "pytest>=9.0.2",
    "pytest-playwright>=0.7.2",
    "ruff>=0.15.5",
]

[tool.pytest.ini_options]
pythonpath = ["."]
